# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentTape API",
    "description": "Read API for the AgentTape live AI-agent index. Every score on the wire is an envelope \u2014 headline + four pillars + manipulation_resistance. There is no headline-only path.",
    "version": "0.0.0"
  },
  "paths": {
    "/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List Agents",
        "operationId": "list_agents_agents_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring match on slug/name/description",
              "title": "Q"
            },
            "description": "Substring match on slug/name/description"
          },
          {
            "name": "tag_kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tag kind (capability/domain/...)",
              "title": "Tag Kind"
            },
            "description": "Filter by tag kind (capability/domain/...)"
          },
          {
            "name": "tag_value",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tag value",
              "title": "Tag Value"
            },
            "description": "Filter by tag value"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(score|discovered|name)$",
              "default": "score",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AgentSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/{slug}": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent",
        "operationId": "get_agent_agents__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/{slug}/signals": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent Signals",
        "operationId": "get_agent_signals_agents__slug__signals_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "sources",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by signal_source values",
              "title": "Sources"
            },
            "description": "Filter by signal_source values"
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(1h|1d|7d|30d|90d|all)$",
              "default": "30d",
              "title": "Window"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "default": 2000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignalSeries"
                  },
                  "title": "Response Get Agent Signals Agents  Slug  Signals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/{slug}/benchmarks": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent Benchmarks",
        "operationId": "get_agent_benchmarks_agents__slug__benchmarks_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BenchmarkResultOut"
                  },
                  "title": "Response Get Agent Benchmarks Agents  Slug  Benchmarks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/{slug}/similar": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Similar Agents",
        "operationId": "get_similar_agents_agents__slug__similar_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimilarAgent"
                  },
                  "title": "Response Get Similar Agents Agents  Slug  Similar Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/indexes/{slug}": {
      "get": {
        "tags": [
          "indexes"
        ],
        "summary": "Get Index",
        "operationId": "get_index_indexes__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/indexes/{slug}/history": {
      "get": {
        "tags": [
          "indexes"
        ],
        "summary": "Get Index History",
        "operationId": "get_index_history_indexes__slug__history_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(1d|7d|30d|90d|all)$",
              "default": "30d",
              "title": "Window"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "default": 2000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IndexSnapshotOut"
                  },
                  "title": "Response Get Index History Indexes  Slug  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/indexes/{slug}/rebalances": {
      "get": {
        "tags": [
          "indexes"
        ],
        "summary": "Get Index Rebalances",
        "operationId": "get_index_rebalances_indexes__slug__rebalances_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RebalanceOut"
                  },
                  "title": "Response Get Index Rebalances Indexes  Slug  Rebalances Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "discovered_via": {
            "type": "string",
            "title": "Discovered Via"
          },
          "discovered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Discovered At"
          },
          "homepage_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Homepage Url"
          },
          "github_repo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Repo"
          },
          "score": {
            "$ref": "#/components/schemas/ScoreEnvelopeOptional"
          },
          "hf_org": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hf Org"
          },
          "hf_model_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hf Model Ids"
          },
          "package_names": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Names"
          },
          "arxiv_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arxiv Ids"
          },
          "eligibility_status": {
            "type": "string",
            "title": "Eligibility Status"
          },
          "eligibility_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eligibility Score"
          },
          "eligibility_reasons": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eligibility Reasons"
          },
          "manipulation_flags": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manipulation Flags"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "discovered_via",
          "discovered_at",
          "homepage_url",
          "github_repo",
          "score",
          "hf_org",
          "hf_model_ids",
          "package_names",
          "arxiv_ids",
          "eligibility_status",
          "eligibility_score",
          "eligibility_reasons",
          "manipulation_flags",
          "tags"
        ],
        "title": "AgentDetail",
        "description": "The full /agents/:slug payload."
      },
...
      "BenchmarkResultOut": {
        "properties": {
          "benchmark_id": {
            "type": "string",
            "format": "uuid",
            "title": "Benchmark Id"
          },
          "benchmark_name": {
            "type": "string",
            "title": "Benchmark Name"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "max_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Score"
          }
        },
        "type": "object",
        "required": [
          "benchmark_id",
          "benchmark_name",
          "captured_at",
          "score",
          "max_score"
        ],
        "title": "BenchmarkResultOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IndexDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "methodology_md": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Methodology Md"
          },
          "rebalance_frequency": {
            "type": "string",
            "title": "Rebalance Frequency"
          },
          "members_count": {
            "type": "integer",
            "title": "Members Count"
          },
          "composite_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Composite Value"
          },
          "constituents": {
            "items": {
              "$ref": "#/components/schemas/IndexConstituent"
            },
            "type": "array",
            "title": "Constituents"
          },
          "last_rebalance_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Rebalance At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "methodology_md",
          "rebalance_frequency",
          "members_count",
          "composite_value",
          "constituents",
          "last_rebalance_at"
        ],
        "title": "IndexDetail"
      },
      "IndexSnapshotOut": {
        "properties": {
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          },
          "composite_value": {
            "type": "number",
            "title": "Composite Value"
          }
        },
        "type": "object",
        "required": [
          "captured_at",
          "composite_value"
        ],
        "title": "IndexSnapshotOut"
      },
...
      "Page_AgentSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[AgentSummary]"
      },
...
      "RebalanceOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "run_at": {
            "type": "string",
            "format": "date-time",
            "title": "Run At"
          },
          "additions": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Additions"
          },
          "removals": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Removals"
          },
          "weight_changes": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Changes"
          },
          "narrative_md": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative Md"
          }
        },
        "type": "object",
        "required": [
          "id",
          "run_at",
          "additions",
          "removals",
          "weight_changes",
          "narrative_md"
        ],
        "title": "RebalanceOut"
      },
...
      "SignalSeries": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/SignalPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "source",
          "points"
        ],
        "title": "SignalSeries"
      },
      "SimilarAgent": {
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/AgentSummary"
          },
          "similarity": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Similarity"
          }
        },
        "type": "object",
        "required": [
          "agent",
          "similarity"
        ],
        "title": "SimilarAgent"
      },
      "TagOut": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "value",
          "display_name"
        ],
        "title": "TagOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}